extracted the pending messages
